codacy improvements
