+++ handlers/sf-contact-merge/src/main/scala/com/gu/sf_contact_merge/Handler.scala
@@ -8,7 +8,7 @@ import com.gu.salesforce.SalesforceAuthenticate.SFAuthConfig
 import com.gu.salesforce.TypesForSFEffectsData.SFContactId
 import com.gu.salesforce.{JsonHttp, SalesforceClient}
 import com.gu.sf_contact_merge.GetSFIdentityIdMoveData.{CanonicalEmail, SFContactIdEmailIdentity}
-import com.gu.sf_contact_merge.SFSteps.SFData
+import com.gu.sf_contact_merge.SFSteps.{GetSfContact, SFData}
 import com.gu.sf_contact_merge.TypeConvert._
 import com.gu.sf_contact_merge.Types.WinningSFContact
 import com.gu.sf_contact_merge.WireRequestToDomainObject.MergeRequest
@@ -17,8 +17,9 @@ import com.gu.sf_contact_merge.getaccounts.GetContacts.AccountId
 import com.gu.sf_contact_merge.getaccounts.GetZuoraContactDetails.{EmailAddress, LastName}
 import com.gu.sf_contact_merge.getaccounts.{GetIdentityAndZuoraEmailsForAccountsSteps, GetZuoraContactDetails}
 import com.gu.sf_contact_merge.getsfcontacts.DedupSfContacts.SFContactsForMerge
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact.SFContact
-import com.gu.sf_contact_merge.getsfcontacts.{DedupSfContacts, GetSfAddressOverride, GetSfContact}
+import com.gu.sf_contact_merge.getsfcontacts.ToSfContactRequest.WireResult
+import com.gu.sf_contact_merge.getsfcontacts.WireContactToSfContact.Types.SFContact
+import com.gu.sf_contact_merge.getsfcontacts.{DedupSfContacts, GetSfAddressOverride, ToSfContactRequest, WireContactToSfContact}
 import com.gu.sf_contact_merge.update.UpdateAccountSFLinks.{CRMAccountId, ZuoraFieldUpdates}
 import com.gu.sf_contact_merge.update.{UpdateAccountSFLinks, UpdateSFContacts, UpdateSalesforceIdentityId}
 import com.gu.sf_contact_merge.validate.AssertSame.{Differing, HasAllowableVariations, HasNoVariations}
@@ -29,8 +30,9 @@ import com.gu.util.config.LoadConfigModule.StringFromS3
 import com.gu.util.config.{LoadConfigModule, Stage}
 import com.gu.util.reader.Types.ApiGatewayOp.{ContinueProcessing, ReturnWithResponse}
 import com.gu.util.reader.Types._
-import com.gu.util.resthttp.LazyClientFailableOp
+import com.gu.util.resthttp.RestOp.HttpOpParseOp
 import com.gu.util.resthttp.Types.ClientFailableOp
+import com.gu.util.resthttp.{HttpOp, LazyClientFailableOp}
 import com.gu.util.zuora.SafeQueryBuilder.MaybeNonEmptyList
 import com.gu.util.zuora.{ZuoraQuery, ZuoraRestConfig, ZuoraRestRequestMaker}
 import okhttp3.{Request, Response}
@@ -59,6 +61,8 @@ object Handler {
 
     } yield Operation.noHealthcheck {
       WireRequestToDomainObject {
+        val sfGet = sfAuth.wrap(JsonHttp.get)
+        val getSfContact = GetSfContact(sfGet.setupRequest(ToSfContactRequest.apply).parse[WireResult].map(WireContactToSfContact.apply))
         DomainSteps(
           ZuoraSteps(
             GetIdentityAndZuoraEmailsForAccountsSteps(zuoraQuerier),
@@ -69,7 +73,7 @@ object Handler {
           SFSteps(
             AssertSame.emailAddress,
             GetSfAddressOverride.apply,
-            GetSfContact(sfAuth.wrap(JsonHttp.get))
+            getSfContact
           )
         )
       }
@@ -112,12 +116,15 @@ object DomainSteps {
 }
 
 object SFSteps {
+
+  case class GetSfContact(apply: HttpOp[SFContactId, SFContact])
+
   def apply(
     validateEmails: AssertSame[EmailAddress],
     getSfAddressOverride: GetSfAddressOverride,
     getSfContact: GetSfContact
   ): SFSteps = { dedupedContactIds =>
-    val sfContacts = dedupedContactIds.map(id => getSfContact.apply(id).map(contact => IdWithContact(id, contact)))
+    val sfContacts = dedupedContactIds.map(id => getSfContact.apply.runRequestLazy(id).map(contact => IdWithContact(id, contact)))
     for {
       maybeSFAddressOverride <- getSfAddressOverride(sfContacts.map(_.value.map(_.contact.SFMaybeAddress)))
         .toApiGatewayOp("get salesforce addresses")
@@ -136,7 +143,9 @@ object SFSteps {
     } yield SFData(sfIdentityIdMoveData, emailOverrides, maybeSFAddressOverride)
   }
 
-  def flattenContactData(sfContacts: SFContactsForMerge[LazyClientFailableOp[IdWithContact]]): ClientFailableOp[List[SFContactIdEmailIdentity]] = {
+  def flattenContactData(
+    sfContacts: SFContactsForMerge[LazyClientFailableOp[IdWithContact]]
+  ): ClientFailableOp[List[SFContactIdEmailIdentity]] = {
     val contactsList = NonEmptyList(sfContacts.winner, sfContacts.others: _*)
     val lazyContactsEmailIdentity = contactsList.map(_.map(idWithContact =>
       SFContactIdEmailIdentity(idWithContact.id, idWithContact.contact.emailIdentity)))
@@ -170,8 +179,10 @@ object ZuoraSteps {
         .toApiGatewayOp("getIdentityAndZuoraEmailsForAccounts")
       _ <- StopIfNoContactsToChange(mergeRequest.winningSFContact.id, zuoraAccountAndEmails.map(_.sfContactId))
       _ <- validateLastNames(zuoraAccountAndEmails.map(_.lastName)) match {
-        case Differing(elements) => ReturnWithResponse(ApiGatewayResponse.notFound(s"those zuora accounts had differing lastname: $elements"))
-        case _ => ContinueProcessing(())
+        case Differing(elements) =>
+          ReturnWithResponse(ApiGatewayResponse.notFound(s"those zuora accounts had differing lastname: $elements"))
+        case _ =>
+          ContinueProcessing(())
       }
       firstNameToUse <- GetFirstNameToUse(mergeRequest.winningSFContact, zuoraAccountAndEmails)
       allSFContactIds = SFContactsForMerge(mergeRequest.winningSFContact.id, zuoraAccountAndEmails.map(_.sfContactId))
