+++ handlers/sf-contact-merge/src/main/scala/com/gu/sf_contact_merge/validate/ValidateNoLosingDigitalVoucher.scala
@@ -1,8 +1,7 @@
 package com.gu.sf_contact_merge.validate
 
 import com.gu.sf_contact_merge.TypeConvert._
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact.IsDigitalVoucherUser
+import com.gu.sf_contact_merge.getsfcontacts.WireContactToSfContact.Types.IsDigitalVoucherUser
 import com.gu.util.apigateway.ApiGatewayResponse
 import com.gu.util.reader.Types.ApiGatewayOp
 import com.gu.util.reader.Types.ApiGatewayOp.{ContinueProcessing, ReturnWithResponse}
@@ -10,7 +9,7 @@ import com.gu.util.resthttp.LazyClientFailableOp
 
 object ValidateNoLosingDigitalVoucher {
 
-  def apply(losingContacts: List[LazyClientFailableOp[GetSfContact.IsDigitalVoucherUser]]): ApiGatewayOp[Unit] =
+  def apply(losingContacts: List[LazyClientFailableOp[IsDigitalVoucherUser]]): ApiGatewayOp[Unit] =
     losingContacts.toStream.map {
       _.value.toApiGatewayOp("get SF address").flatMap {
         case IsDigitalVoucherUser(true) => ReturnWithResponse(ApiGatewayResponse.notFound(
