+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/getsfcontacts/GetSfAddressEffectsTest.scala
@@ -1,13 +1,14 @@
 package com.gu.sf_contact_merge.getsfcontacts
 
 import com.gu.effects.{GetFromS3, RawEffects}
-import com.gu.salesforce.{JsonHttp, SalesforceClient}
 import com.gu.salesforce.SalesforceAuthenticate.SFAuthConfig
 import com.gu.salesforce.dev.SFEffectsData
+import com.gu.salesforce.{JsonHttp, SalesforceClient}
 import com.gu.sf_contact_merge.Types.IdentityId
 import com.gu.sf_contact_merge.getaccounts.GetZuoraContactDetails.EmailAddress
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact.SFAddressFields._
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact.{EmailIdentity, IsDigitalVoucherUser, SFAddress, SFContact, UsableContactAddress}
+import com.gu.sf_contact_merge.getsfcontacts.ToSfContactRequest.WireResult
+import com.gu.sf_contact_merge.getsfcontacts.WireContactToSfContact.Types._
+import com.gu.util.resthttp.RestOp._
 import com.gu.test.EffectsTest
 import com.gu.util.config.{LoadConfigModule, Stage}
 import org.scalatest.{FlatSpec, Matchers}
@@ -23,9 +24,8 @@ class GetSfAddressEffectsTest extends FlatSpec with Matchers {
       sfConfig <- LoadConfigModule(Stage("DEV"), GetFromS3.fetchString)[SFAuthConfig]
       response = RawEffects.response
       sfAuth <- SalesforceClient(response, sfConfig).value.toDisjunction
-      get = sfAuth.wrap(JsonHttp.get)
-      getSfAddress = GetSfContact(get)
-      address <- getSfAddress.apply(testContact).value.toDisjunction
+      getSfContact = sfAuth.wrap(JsonHttp.get).setupRequest(ToSfContactRequest.apply).parse[WireResult].map(WireContactToSfContact.apply)
+      address <- getSfContact.runRequest(testContact).toDisjunction
     } yield address
 
     val expected = SFContact(
