+++ handlers/sf-contact-merge/src/test/scala/com/gu/sf_contact_merge/getsfcontacts/WireContactToSfContactTest.scala
@@ -1,19 +1,12 @@
 package com.gu.sf_contact_merge.getsfcontacts
 
-import com.gu.salesforce.TypesForSFEffectsData.SFContactId
 import com.gu.sf_contact_merge.Types.IdentityId
 import com.gu.sf_contact_merge.getaccounts.GetZuoraContactDetails.EmailAddress
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact.SFAddressFields._
-import com.gu.sf_contact_merge.getsfcontacts.GetSfContact.{EmailIdentity, IsDigitalVoucherUser, SFAddress, SFMaybeAddress, UnusableContactAddress, UsableContactAddress, WireResult}
-import com.gu.util.resthttp.RestRequestMaker.{GetRequest, RelativePath}
+import com.gu.sf_contact_merge.getsfcontacts.ToSfContactRequest.WireResult
+import com.gu.sf_contact_merge.getsfcontacts.WireContactToSfContact.Types._
 import org.scalatest.{FlatSpec, Matchers}
 
-class GetSfAddressTest extends FlatSpec with Matchers {
-
-  "toRequest" should "compose a correct GET request" in {
-    val actual = GetSfContact.toRequest(SFContactId("testcont"))
-    actual should be(GetRequest(RelativePath("/services/data/v43.0/sobjects/Contact/testcont")))
-  }
+class WireContactToSfContactTest extends FlatSpec with Matchers {
 
   "toMaybeAddress" should "return some if all are set" in {
     val wireResult = WireResult(
@@ -27,7 +20,7 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: SFMaybeAddress = GetSfContact.toMaybeAddress(wireResult)
+    val actual: SFMaybeAddress = WireContactToSfContact.toMaybeAddress(wireResult)
 
     val expected = SFAddress(
       SFStreet("street1"),
@@ -53,7 +46,7 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: SFMaybeAddress = GetSfContact.toMaybeAddress(wireResult)
+    val actual: SFMaybeAddress = WireContactToSfContact.toMaybeAddress(wireResult)
 
     val expected = SFAddress(
       SFStreet("street1"),
@@ -79,7 +72,7 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: SFMaybeAddress = GetSfContact.toMaybeAddress(wireResult)
+    val actual: SFMaybeAddress = WireContactToSfContact.toMaybeAddress(wireResult)
 
     actual should be(UnusableContactAddress)
   }
@@ -96,7 +89,7 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: SFMaybeAddress = GetSfContact.toMaybeAddress(wireResult)
+    val actual: SFMaybeAddress = WireContactToSfContact.toMaybeAddress(wireResult)
 
     actual should be(UnusableContactAddress)
   }
@@ -113,7 +106,7 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: SFMaybeAddress = GetSfContact.toMaybeAddress(wireResult)
+    val actual: SFMaybeAddress = WireContactToSfContact.toMaybeAddress(wireResult)
 
     actual should be(UnusableContactAddress)
   }
@@ -130,7 +123,7 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: SFMaybeAddress = GetSfContact.toMaybeAddress(wireResult)
+    val actual: SFMaybeAddress = WireContactToSfContact.toMaybeAddress(wireResult)
 
     actual should be(UnusableContactAddress)
   }
@@ -147,11 +140,11 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = Some("1234")
     )
-    val actual: GetSfContact.IsDigitalVoucherUser = GetSfContact.toResponse(wireResult).isDigitalVoucherUser
+    val actual: IsDigitalVoucherUser = WireContactToSfContact(wireResult).isDigitalVoucherUser
 
     actual should be(IsDigitalVoucherUser(false))
 
-    val actual2: GetSfContact.EmailIdentity = GetSfContact.toResponse(wireResult).emailIdentity
+    val actual2: EmailIdentity = WireContactToSfContact(wireResult).emailIdentity
 
     actual2 should be(EmailIdentity(EmailAddress("a@b.com"), Some(IdentityId("1234"))))
   }
@@ -168,10 +161,10 @@ class GetSfAddressTest extends FlatSpec with Matchers {
       Email = "a@b.com",
       IdentityID__c = None
     )
-    val actual: GetSfContact.IsDigitalVoucherUser = GetSfContact.toResponse(wireResult).isDigitalVoucherUser
+    val actual: IsDigitalVoucherUser = WireContactToSfContact(wireResult).isDigitalVoucherUser
 
     actual should be(IsDigitalVoucherUser(true))
-    val actual2: GetSfContact.EmailIdentity = GetSfContact.toResponse(wireResult).emailIdentity
+    val actual2: EmailIdentity = WireContactToSfContact(wireResult).emailIdentity
 
     actual2 should be(EmailIdentity(EmailAddress("a@b.com"), None))
   }
